+++ QMA.ViewModel/Observables/Season/ObservableTeamMember.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Toolkit.Mvvm.ComponentModel;
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text;
@@ -11,7 +12,7 @@ namespace QMA.ViewModel.Observables.Season
 {
     public class ObservableTeamMember : ObservableValidator
     {
-        public ObservableTeamMember(bool persisted, string quizzerId, string quizzerName)
+        public ObservableTeamMember(bool persisted, string teamMemberId, string quizzerId, string quizzerName)
         {
             if (quizzerId == null)
             {
@@ -22,6 +23,8 @@ namespace QMA.ViewModel.Observables.Season
                 throw new ArgumentNullException(nameof(quizzerName));
             }
 
+            _teamMemberId = teamMemberId;
+
             _quizzerId = quizzerId;
 
             _quizzerName = quizzerName;
@@ -29,6 +32,13 @@ namespace QMA.ViewModel.Observables.Season
             Persisted = persisted;
 
             ValidateAllProperties();
+        }
+
+        private string _teamMemberId;
+        public string TeamMemberId
+        {
+            get => _teamMemberId;
+            set => SetProperty(ref _teamMemberId, value);
         }
 
         private string _quizzerId;
@@ -41,6 +51,12 @@ namespace QMA.ViewModel.Observables.Season
         public string QuizzerName
         {
             get => _quizzerName;
+        }
+
+        private ObservableCollection<ObservableAssignedQuestion> _assignedQuestions = new ObservableCollection<ObservableAssignedQuestion>();
+        public ObservableCollection<ObservableAssignedQuestion> AssignedQuestions
+        {
+            get => _assignedQuestions;
         }
 
         private bool _isMember;