+++ CHANGELOG.md
@@ -53,3 +53,4 @@
 
 ## 0.5.0
 - Implement assignment of questions to quizzers on a team
+- Fix bug with generating primary keys on imported questions